+++ packages/security/toolsets.json
@@ -1,0 +1,5 @@
+{
+  "@valos/type-library": {
+    "inUse": "always"
+  }
+}
