+++ packages/sourcerer/package.json
@@ -16,8 +16,8 @@
   "main": "index.js",
   "dependencies": {
     "@valos/script": "^0.35.0-rc.35",
-    "graphql": "^0.6.2",
-    "jssha": "^2.3.1"
+    "@valos/security": "^0.35.0-rc.35",
+    "graphql": "^0.6.2"
   },
   "homepage": "https://valospace.org/sourcerer"
 }
