+++ packages/sourcerer/tools/event-version-0.2/createResourceId0Dot2.js
@@ -1,6 +1,6 @@
 // @flow
 
-import hashV240 from "~/sourcerer/tools/hashV240";
+import { hashV240 } from "~/security/hash";
 
 export default function createResourceId0Dot2 (commandId: string, chronicleURI: string,
     intraEventIndex: ?number) {
