+++ packages/tools/package.json
@@ -17,9 +17,9 @@
     "@babel/runtime": ">7.0.0-rc",
     "@babel/runtime-corejs2": ">7.0.0-rc",
     "@valos/gateway-api": "^0.35.0-rc.35",
+    "@valos/security": "^0.35.0-rc.35",
     "fake-indexeddb": "^2.0.3",
     "immutable": "~3.7.4",
-    "jssha": "^2.3.1",
     "object-hash": "^1.1.3",
     "uuid": "^3.2.1"
   }
