+++ packages/web-spindle/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@valos/type-spindle": "^0.35.0-rc.35",
     "@valos/type-toolset": "^0.35.0-rc.35",
+    "@valos/security": "^0.35.0-rc.35",
     "fastify": "^2.3.0",
     "fastify-cookie": "^3.0.2",
     "fastify-swagger": "^2.3.2"
