+++ packages/web-spindle/tools/session.js
@@ -1,21 +1,13 @@
 // @flow
 
-// import buffer from "buffer";
-import crypto from "crypto";
-
 import { vRef } from "~/raem/VRL";
 import { VALEK, Vrapper } from "~/engine";
 
 import { dumpObject } from "~/tools";
-import {
-  base64URLDecode, base64URLEncode, base64URLFromBuffer, byteArrayFromBase64URL,
-  base64URLFromBase64,
-} from "~/gateway-api/base64";
 
-const _normalizeAlg = Object.assign(Object.create(null), {
-  A256GCM: "aes-256-gcm",
-  "aes-256-gcm": "aes-256-gcm",
-});
+import {
+  generateBurlaesgIV, burlaesgEncode, burlaesgDecode, hs256JWTEncode, hs256JWTDecode,
+} from "~/security/tokens";
 
 const _isReadOnlyMethod = {
   CONNECT: true,
@@ -180,72 +172,3 @@ const accessFields = {
 };
 const toRIGHTSFields = VALEK.relations("RIGHTS").map(VALEK.select(accessFields));
 // const toPERMISSIONSFields = VALEK.relations("PERMISSIONS").map(VALEK.select(accessFields));
-
-export function generateBurlaesgIV () {
-  // TODO(iridian, 2020-01): This should be randomized on start and
-  // then 'one'-incremented for every request
-  return new Uint8Array(crypto.randomBytes(12));
-}
-
-export function burlaesgEncode (payload: any, key: string, iv: Uint8Array) {
-  const cipher = crypto.createCipheriv("aes-256-gcm", _to32BytesNullPad(key), iv);
-  let enc = cipher.update(base64URLEncode(JSON.stringify(payload)), "utf8", "base64");
-  enc += cipher.final("base64");
-  return base64URLEncode(JSON.stringify({
-    iv: base64URLFromBuffer(iv),
-    tag: base64URLFromBuffer(cipher.getAuthTag()),
-    alg: "A256GCM",
-    ciphertext: enc,
-  }));
-}
-
-// &/&¤# kludge
-function _to32BytesNullPad (key: string) {
-  return key + String.fromCharCode(0).repeat(Math.max(0, 32 - key.length));
-}
-
-export function burlaesgDecode (base64URLCode: string, key: string) {
-  const code = base64URLDecode(base64URLCode);
-  const { iv: bIV, tag: bTag, alg, ciphertext } = JSON.parse(code);
-  const iv = byteArrayFromBase64URL(bIV);
-  const tag = byteArrayFromBase64URL(bTag);
-  const normalizedAlg = _normalizeAlg[alg];
-  if (!normalizedAlg) throw new Error("Unrecognized session encryption algorithm");
-  // console.log("base64URLCode:", base64URLCode);
-  // console.log("code:", code, "\n\tiv, tag, alg, cipher:", iv, tag, alg, ciphertext);
-  const decipher = crypto.createDecipheriv(normalizedAlg, _to32BytesNullPad(key), iv);
-  decipher.setAuthTag(tag);
-  const dec = decipher.update(ciphertext, "base64", "utf8");
-  // console.log("pre-final dec:", dec);
-  return { iv, alg, payload: JSON.parse(base64URLDecode(dec + decipher.final("utf8"))) };
-}
-
-export function hs256JWTEncode (payload: string, key: string,
-    header = { alg: "HS256", typ: "JWT" }) {
-  if (!(key.length >= 16)) throw new Error("Invalid key with length < 16");
-  const base64HeaderAndPayload =
-      `${base64URLEncode(JSON.stringify(header))}.${base64URLEncode(JSON.stringify(payload))}`;
-  const signer = crypto.createHmac("sha256", key);
-  signer.update(base64HeaderAndPayload);
-  return `${base64HeaderAndPayload}.${base64URLFromBase64(signer.digest("base64"))}`;
-}
-
-export function hs256JWTDecode (jwt: string, validationKey: ?string) {
-  const parts = jwt.split(".");
-  const ret = {
-    header: JSON.parse(base64URLDecode(parts[0])),
-    payload: JSON.parse(base64URLDecode(parts[1])),
-    signature: parts[2],
-  };
-  if (ret.header.alg !== "HS256") {
-    throw new Error(`Cannot decode non-HS256 JWT with alg = '${ret.header.alg}'`);
-  }
-  if (validationKey) {
-    const signer = crypto.createHmac("sha256", validationKey);
-    signer.update(jwt.substr(0, parts[0].length + 1 + parts[1].length));
-    if (base64URLFromBase64(signer.digest("base64")) !== parts[2]) {
-      throw new Error("HS256 JWT signature mismatch");
-    }
-  }
-  return ret;
-}
